Massive UI bugfix
There is still a problem, that a server only notifies the clients about the new map, once it finished loading it, so they client can't start loading it earlier.
It is a problem, because the countdown start once the server is loaded, so there is a possibility, that the clients won't load in  time, causing them to crash.